# Conflicts:
#   LiDARC-backend/src/main/java/com/example/lidarcbackend/api/comparison/ComparisonController.java
#   LiDARC-frontend/src/app/app.routes.ts
#   LiDARC-frontend/src/app/components/navbar/navbar.html
#   LiDARC-frontend/src/app/service/comparison.service.ts

+++ LiDARC-backend/src/main/java/com/example/lidarcbackend/repository/ComparisonFileRepository.java
@@ -3,10 +3,9 @@ package com.example.lidarcbackend.repository;
 import com.example.lidarcbackend.model.entity.ComparisonFile;
 import com.example.lidarcbackend.model.entity.ComparisonFilePK;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 import java.util.Optional;
@@ -14,7 +13,7 @@ import java.util.Optional;
 @Repository
 public interface ComparisonFileRepository extends JpaRepository<ComparisonFile, ComparisonFilePK> {
     @Query("SELECT cfs FROM ComparisonFile cfs WHERE cfs.comparisonId = :comparisonId AND cfs.fileId = :fileId")
-    Optional<ComparisonFile> findComparisonFiles(int comparisonId, int fileId);
+    Optional<ComparisonFile> findComparisonFiles(Long comparisonId, Long fileId);
 
     @Query("SELECT cf.fileId FROM ComparisonFile cf WHERE cf.comparisonId = :comparisonId")
     List<Long> getComparisonFilesByComparisonId(Long comparisonId);
